Use "module" instead of "main" when exposing ESM.
In order for this package to work out of the box with webpack, we need to either name the index.js "index.mjs" (and not have a file extension, e.g. "esm/index") or use the "module" field in package.json. Otherwise it's not easily possible to load this through webpack without transpiling it.